Move the grid class to its own file
